third_party/WebKit/LayoutTests/bluetooth/server/device-same-object.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../resources/mojo-helpers.js"></script>
<script>
'use strict';
const test_desc = "[SameObject] test for BluetoothRemoteGATTServer's device.";
let device, fake_peripheral;
promise_test(() => getDiscoveredHealthThermometerDevice()
  .then(_ => ({device, fake_peripheral} = _))
  .then(() => fake_peripheral.setNextGATTConnectionResponse({
    code: HCI_SUCCESS,
  }))
  .then(() => device.gatt.connect())
  .then(gatt => assert_equals(gatt.device, gatt.device))
  .then(() => navigator.bluetooth.test.allResponsesConsumed())
  .then(consumed => assert_true(consumed)), test_desc);
</script>
